Project1/requirements.txt
torch
pandas
numpy
scikit-learn
matplotlib
seaborn
requests
beautifulsoup4
tldextract
python-whois
joblib
python-dotenv
